requirements.txt
Hug
Pillow
zappa
